refactor: move service account util from core
